move useTheme hook
